[App] Widget Data and Layout
