--- app-ios/today/TodayViewController.swift
@@ -1,34 +1,0 @@
-//
-//  TodayViewController.swift
-//  today
-//
-//  Created by Martin Púčik on 02/02/2018.
-//  Copyright © 2018 The Cave. All rights reserved.
-//
-
-import UIKit
-import NotificationCenter
-
-class TodayViewController: UIViewController, NCWidgetProviding {
-        
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view from its nib.
-    }
-    
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-    
-    func widgetPerformUpdate(completionHandler: (@escaping (NCUpdateResult) -> Void)) {
-        // Perform any setup necessary in order to update the view.
-        
-        // If an error is encountered, use NCUpdateResult.Failed
-        // If there's no update required, use NCUpdateResult.NoData
-        // If there's an update, use NCUpdateResult.NewData
-        
-        completionHandler(NCUpdateResult.newData)
-    }
-    
-}
