+++ app-ios/wifilamp/Views/DeviceDetail/DeviceDetailVM.swift
@@ -65,10 +65,10 @@ class DeviceDetailVM {
     }
     
     func updateColor(color: UIColor) {
-        // swiftlint:disable identifier_name
         var r: CGFloat = 0
         var g: CGFloat = 0
         var b: CGFloat = 0
+        // swiftlint:enable identifier_name
         color.getRed(&r, green: &g, blue: &b, alpha: nil)
         
         let roundR =  round(r * 255.0)
